requirements.txt
alabaster==0.7.13
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
astroid==2.15.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==1.6.0
Babel==2.12.1
backcall==0.2.0
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coverage==7.2.2
crispy-bootstrap5==0.7
cryptography==40.0.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.6
distlib==0.3.6
dj-database-url==0.5.0
Django==4.1.8
django-allauth==0.54.0
django-auditlog==2.2.2
django-coverage-plugin==3.0.0
django-crispy-forms==2.0
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-extensions==3.2.1
django-import-export==3.1.0
django-model-utils==4.3.1
django-redis==5.2.0
django-stubs==1.16.0
django-stubs-ext==0.8.0
djangorestframework==3.14.0
docutils==0.19
environ==1.0
et-xmlfile==1.1.0
executing==1.2.0
factory-boy==3.2.1
Faker==18.4.0
filelock==3.11.0
flake8==6.0.0
flake8-isort==6.0.0
gunicorn==20.1.0
identify==2.5.22
idna==3.4
imagesize==1.4.1
import-export==0.3.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.12.0
isort==5.12.0
jedi==0.18.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
livereload==2.6.3
MarkupPy==1.14
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy==1.1.1
mypy-extensions==1.0.0
nodeenv==1.7.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.0
parso==0.8.3
pathspec==0.11.1
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
pluggy==1.0.0
pre-commit==3.2.2
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.14.0
PyJWT==2.6.0
pylint==2.17.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytest==7.2.2
pytest-django==4.5.2
pytest-sugar==0.9.6
python-dateutil==2.8.2
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
redis==4.5.4
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==6.1.3
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.3
stack-data==0.6.2
tablib==3.4.0
termcolor==2.2.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.7
tornado==6.2
traitlets==5.9.0
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.9
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.21.0
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
whitenoise==6.4.0
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0
